+++ services/madoc-ts/src/extensions/capture-models/ProjectExplorer/ProjectExplorer.editor.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const ProjectExplorerEditor = () => {
+  return null;
+};
+
+export default ProjectExplorerEditor;