+++ services/madoc-ts/src/extensions/capture-models/ProjectExplorer/ProjectExplorer.preview.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { LocaleString } from '../../../frontend/shared/components/LocaleString';
 import { Heading5 } from '../../../frontend/shared/typography/Heading5';
-import { ProjectSelectorProps } from './ProjectSelector';
+import { ProjectExplorerProps } from './ProjectExplorer';
 
-export const ProjectSelectorPreview: React.FC<ProjectSelectorProps> = props => {
+export const ProjectExplorerPreview: React.FC<ProjectExplorerProps> = props => {
   if (props.value) {
     return <Heading5 as={LocaleString}>{props.value.label}</Heading5>;
   }