+++ services/madoc-ts/src/extensions/capture-models/ProjectSelector/ProjectSelector.editor.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const ProjectSelectorEditor = () => {
-  return null;
-};
-
-export default ProjectSelectorEditor;